refactor: 사용자 키워드 통계 API를 users.py로 이동 및 URI 개선
- GET /users/{user_id}/keyword-stats 엔드포인트를 ratings.py에서 users.py로 이동
- REST 구조를 개선하기 위해 URI를 /users/{user_id}/stats/keywords로 변경
- 관련 import 정리 및 코드 포맷 수정
- 기능은 동일하게 유지하면서 코드 구조를 더 명확하게 개선
