+++ app/api/endpoints/ratings.py
@@ -18,7 +18,6 @@ from app.schemas.responses import (
     RatingWithReviewerResponse,
     ReviewableUserResponse,
     ReviewableUsersResponse,
-    UserKeywordStatsResponse,
 )
 
 # 게시물 관련 평가 API 라우터
@@ -461,38 +460,6 @@ async def get_rating_keywords(
     ]
 
     return RatingKeywordListResponse(keywords=keywords_data)
-
-
-@router.get(
-    "/users/{user_id}/keyword-stats",
-    response_model=UserKeywordStatsResponse,
-    description="사용자 키워드 통계 조회",
-)
-async def get_user_keyword_stats(
-    user_id: str,
-    session: AsyncSession = Depends(get_async_session),
-) -> UserKeywordStatsResponse:
-    """사용자의 키워드 통계를 조회합니다."""
-
-    # 사용자 존재 확인
-    user_query = select(User).where(User.id == user_id)
-    user_result = await session.execute(user_query)
-    user = user_result.scalar_one_or_none()
-
-    if not user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="사용자를 찾을 수 없습니다.",
-        )
-
-    # TODO: UserKeywordStats 모델이 구현되면 실제 통계 데이터 조회
-    # 현재는 임시 데이터 반환
-    return UserKeywordStatsResponse(
-        user_id=user_id,
-        keyword_code="temp",
-        count=0,
-        last_updated=datetime.utcnow(),
-    )
 
 
 @router.get(
